+++ src/main/java/com/triprint/backend/domain/user/controller/MyProfileController.java
@@ -2,11 +2,9 @@ package com.triprint.backend.domain.user.controller;
 
 import javax.servlet.http.HttpServletRequest;
 
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -16,7 +14,6 @@ import org.springframework.web.multipart.MultipartFile;
 
 import com.triprint.backend.domain.auth.security.CurrentUser;
 import com.triprint.backend.domain.auth.security.UserPrincipal;
-import com.triprint.backend.domain.user.dto.FollowUserRequest;
 import com.triprint.backend.domain.user.dto.MyProfileImgResponse;
 import com.triprint.backend.domain.user.dto.MyProfileRequest;
 import com.triprint.backend.domain.user.dto.MyProfileResponse;
@@ -53,12 +50,5 @@ public class MyProfileController {
 	) {
 		return ResponseEntity.ok(
 			myProfileService.updateMyProfileImg((Long)request.getAttribute("userId"), multipartFile));
-	}
-
-	@PostMapping("/follow")
-	@PreAuthorize("hasRole('ROLE_USER')")
-	ResponseEntity<Object> followUser(@CurrentUser UserPrincipal userPrincipal, FollowUserRequest followUserRequest) {
-		myProfileService.followUser(userPrincipal.getId(), followUserRequest.getId());
-		return ResponseEntity.status(HttpStatus.NO_CONTENT).body(null);
 	}
 }
